+++ android/app/src/main/java/poke/rogue/helper/presentation/battle/selection/BattleSelectionPagerAdapter.kt
@@ -3,6 +3,8 @@ package poke.rogue.helper.presentation.battle.selection
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import poke.rogue.helper.presentation.battle.selection.pokemon.PokemonSelectionFragment
+import poke.rogue.helper.presentation.battle.selection.skill.SkillSelectionFragment
 
 class BattleSelectionPagerAdapter(fragmentActivity: FragmentActivity) :
     FragmentStateAdapter(fragmentActivity) {
