+++ android/app/src/main/res/drawable/bg_before_button.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/poke_grey_60" />
+    <corners android:radius="4dp" />
+</shape>
