+++ android/app/src/main/res/drawable/bg_search_view.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <solid android:color="@color/poke_grey_80" />
-    <corners android:radius="30dp" />
+    <corners android:radius="16dp" />
     <padding
         android:left="8dp"
         android:right="4dp" />
